requirements.txt
flet>=0.21.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
PyPDF2>=3.0.0 